+++ requirements.txt
@@ -1,3 +1,4 @@
+matplotlib
 scipy >= 1.4
 sklearn
 statsmodels
