+++ src/acsbm/estimation.py
@@ -1,267 +1,15 @@
 from collections import Counter, defaultdict
 from dataclasses import dataclass
-import functools
-import itertools
-from typing import Any, Iterable
+from typing import Any
 
 import numpy as np
-from scipy import linalg, optimize, sparse, spatial, stats
-from sklearn import metrics
+from scipy import linalg, optimize, sparse, spatial
 from sklearn.mixture import GaussianMixture
 import statsmodels.genmod.families as sm_families
 from statsmodels.genmod.generalized_linear_model import GLM as sm_GLM
 
-
-def kron_combine(A: np.ndarray, B: np.ndarray) -> np.ndarray:
-    """
-    kron(A, 11^T) + kron(11^T, B)
-    Note: This is not the same as a Kronecker sum, which uses identities in place of the ones matrices.
-    """
-    return np.kron(A, np.ones(B.shape)) + np.kron(np.ones(A.shape), B)
-
-def tuple_id(t: Iterable[int], levels: Iterable[int]) -> int:
-    """
-    A bijection that maps a tuple of ints to a scalar integer
-    Example with levels = (2, 3):
-    (0, 0) -> 0
-    (0, 1) -> 1
-    (0, 2) -> 2
-    (1, 0) -> 3
-    (1, 1) -> 4
-    (1, 2) -> 5
-    """
-    values = list(t)
-    multiplier = 0
-    for i in reversed(range(len(values))):
-        if multiplier > 0:
-            values[i] *= multiplier
-        multiplier += levels[i]
-    
-    return sum(values)
-
-def from_tuple_id(id: int, levels: Iterable[int]) -> tuple[int]:
-    """
-    Inverse of tuple_id
-    """
-    entries = []
-    for i in reversed(range(len(levels))):
-        remainder = id % levels[i]
-        entries.append(remainder)
-        id = id // levels[i]
-    return tuple(reversed(entries))
-
-
-@dataclass
-class Covariate:
-    beta_matrix: np.ndarray # L x L matrix of pairwise group effects, usually diagonal
-
-    @classmethod
-    def simple(cls, beta: float, L: int):
-        return Covariate(beta_matrix=beta * np.eye(L))
-
-    @property
-    def n_levels(self) -> int:
-        return self.beta_matrix.shape[0]
-    
-    def combine(self, other):
-        return Covariate(beta_matrix=kron_combine(self.beta_matrix, other.beta_matrix))
-
-
-@dataclass
-class LinkFunction():
-    """
-    Wrapper around statsmodels link functions.
-    Why? Not sure.
-    """
-    _statsmodels_link: sm_families.family.Family
-
-    def inverse(self, x: np.ndarray) -> np.ndarray:
-        return self._statsmodels_link.inverse(x)
-    
-    @classmethod
-    def identity(cls):
-        return LinkFunction(sm_families.links.identity())
-    
-    @classmethod
-    def log(cls):
-        return LinkFunction(sm_families.links.log())
-    
-    @classmethod
-    def logit(cls):
-        return LinkFunction(sm_families.links.logit())
-    
-    @classmethod
-    def probit(cls):
-        return LinkFunction(sm_families.links.probit())
-
-
-@dataclass
-class MultiCovariateModel:
-    link: LinkFunction
-    B: np.ndarray # K x K matrix of base edge probabilities
-    covariates: list[Covariate]
-
-    @property
-    def n_communities(self):
-        return self.B.shape[0]
-
-    def flatten_covariate(self):
-        """
-        Return a single covariate whose beta_matrix reflects the combined effects of all covariates
-        """
-        return functools.reduce(lambda x, y: x.combine(y), self.covariates)
-
-    def flatten_Z(self, Z: np.ndarray) -> np.ndarray:
-        """
-        Map each covariate combination to a unique scalar
-        Essentially tuple_id, but implemented in numpy instead.
-        """
-        levels = [c.n_levels for c in self.covariates]
-        multiplier = 0
-        Z2 = Z.copy()
-        for i in reversed(range(Z2.shape[1])):
-            if multiplier > 0:
-                Z2[:,i] *= multiplier
-            multiplier += levels[i]
-
-        return Z2.sum(axis=1)
-    
-    def B_tilde(self):
-        flattened = self.flatten_covariate()
-        M = kron_combine(self.B, flattened.beta_matrix)
-        return self.link.inverse(M)
-
-
-class NodeDataDistribution:
-    """
-    A distribution on latent communities (theta) and covariates (Z)
-    pmf is a K x L_1 x L_2 x ... x L_P array of probabilities, where
-    pmf[k, l1, ... lP] = P(theta = k, Z_1 = l1, ... Z_P = lP)
-    """
-    
-    def __init__(self, pmf: np.ndarray):
-        self.pmf = pmf / np.sum(pmf) # normalize
-
-    def draw(self, n: int, rng: np.random.Generator = None) -> np.ndarray:
-        if rng is None:
-            rng = np.random.default_rng()
-
-        all_indices = list(np.ndindex(*self.pmf.shape))
-        data = rng.choice(all_indices, n, p = self.pmf.reshape(len(all_indices),))
-        return data[:,0], data[:,1:] # theta, Z
-
-    @classmethod
-    def _model_shape(cls, model: MultiCovariateModel) -> tuple[int]:
-        return tuple([model.n_communities] + [c.n_levels for c in model.covariates])
-
-    @classmethod
-    def uniform_for_model(cls, model: MultiCovariateModel):
-        return NodeDataDistribution(np.ones(cls._model_shape(model)))
-
-    @classmethod
-    def random_for_model(cls, model: MultiCovariateModel):
-        """
-        An arbitrary random, non-uniform distribution
-        """
-        return NodeDataDistribution(np.random.default_rng().beta(2, 3, cls._model_shape(model)) + 0.05)
-
-
-class NetworkTooSmallError(RuntimeError):
-    pass
-
-
-# pmf = np.array([[
-#     [1, 3],
-#     [0, 6]
-# ],
-# [
-#     [0, 5],
-#     [5, 0]
-# ]])
-# empirical = np.zeros_like(pmf)
-# gen = multicovariates.NodeDataDistribution(pmf)
-# n = 1000
-# for x in gen.draw(n):
-#     empirical[tuple(x)] += 1
-# empirical / n
-
-@dataclass
-class GeneratedNetwork:
-    A: Any # scipy.sparse
-    theta: np.ndarray
-    Z: np.ndarray
-    Z_tilde: np.ndarray
-    block_sizes: list[int]
-    model: MultiCovariateModel
-
-    @property
-    def n_blocks(self) -> int:
-        return len(self.block_sizes)
-
-    @property
-    def n_nodes(self) -> int:
-        return self.A.shape[0]
-    
-    @property
-    def n_edges(self) -> int:
-        return self.A.count_nonzero()
-
-
-def generate_sparse_block(size, prob, symmetric=False):
-    """
-    Generates a random block of binary entries where each entry is 1 w.p. prob
-    If symmetric=True, returns a symmetric block with a zero on the diagonal.
-    """
-    density = stats.binom.rvs(size[0] * size[1], prob, size=1).item() / (size[0] * size[1])
-    m = sparse.random(size[0], size[1], density)
-    m.data[:] = 1
-    
-    if symmetric:
-        if size[0] != size[1]:
-            raise RuntimeError("symmetric matrix must be square")
-        m = sparse.triu(m, k=1) + sparse.triu(m, k=1).transpose()
-    
-    return m
-
-def generate_sparse_sbm(block_sizes, block_probs):
-    """
-    Generate a stochastic block model using fixed block sizes and connectivity matrix
-    """
-    k = len(block_sizes)
-    blocks = [[None for i in range(k)] for j in range(k)]
-    
-    for i in range(k):
-        for j in range(i, k):
-            blocks[i][j] = generate_sparse_block(
-                (block_sizes[i], block_sizes[j]),
-                block_probs[i,j],
-                symmetric=(i == j)
-            )
-            if i < j:
-                blocks[j][i] = blocks[i][j].transpose()
-    
-    return sparse.bmat(blocks)
-
-def generate_network(model: MultiCovariateModel, ndd: NodeDataDistribution, n: int):
-    theta, Z = ndd.draw(n)
-    Z_tilde = model.flatten_Z(Z)
-    L_tilde = np.prod([c.n_levels for c in model.covariates])
-    theta_tilde = L_tilde * theta + Z_tilde # same as applying tuple_id(...) over rows
-    counts = Counter(theta_tilde)
-    block_sizes = [counts[i] for i in range(model.n_communities * L_tilde)]
-
-    # This could probably be handled more gracefully...
-    if len(counts.keys()) < model.n_communities * L_tilde:
-        raise NetworkTooSmallError("Generated network does not have a node of every type. Consider using a larger n.")
-
-    return GeneratedNetwork(
-        A=generate_sparse_sbm(block_sizes, model.B_tilde()),
-        theta=theta[np.argsort(theta_tilde)],
-        Z=Z[np.argsort(theta_tilde), :],
-        Z_tilde=Z_tilde[np.argsort(theta_tilde)],
-        block_sizes=block_sizes,
-        model=model
-    )
+from . import generation
+from .utils import from_tuple_id, tuple_id
 
 
 @dataclass
@@ -350,7 +98,7 @@ def scaled_embeddings(M: Any, d: int = None) -> np.ndarray:
         l, U = sparse.linalg.eigsh(M, d)
     return np.multiply(np.abs(l)**0.5, U)
 
-def initial_cluster(net: GeneratedNetwork, k: int, d: int) -> NetworkInitialClusterResult:
+def initial_cluster(net: generation.GeneratedNetwork, k: int, d: int) -> NetworkInitialClusterResult:
     """
     ACSBM Clustering Part 1:
     Recover theta_tilde (up to a permutation)
@@ -392,7 +140,7 @@ def initial_cluster(net: GeneratedNetwork, k: int, d: int) -> NetworkInitialClus
         theta_tilde=theta_tilde
     )
 
-def reconcile_clusters(net: GeneratedNetwork, initial_cluster: NetworkInitialClusterResult) -> NetworkClusterResult:
+def reconcile_clusters(net: generation.GeneratedNetwork, initial_cluster: NetworkInitialClusterResult) -> NetworkClusterResult:
     """
     ACSBM Clustering Parts 2 and 3:    
     Given an initial clustering (which separately clusters over the partitioned network),
@@ -475,7 +223,7 @@ def invert_permutation(p):
     """
     return [x[0] for x in sorted(enumerate(p), key=lambda x: x[1])]
 
-def estimate(net: GeneratedNetwork, cluster_result: NetworkClusterResult) -> NetworkEstimationResult:
+def estimate(net: generation.GeneratedNetwork, cluster_result: NetworkClusterResult) -> NetworkEstimationResult:
     """
     Estimate B matrix and coefficients of ACSBM model (assuming simple covariates only) via GLM
     """
@@ -534,18 +282,3 @@ def estimate(net: GeneratedNetwork, cluster_result: NetworkClusterResult) -> Net
         B_hat=B_hat,
         coefficients=coef
     )
-
-def label_accuracy(labels, truth):
-    """
-    max(accuracy) over the set of all label permutations
-    truth should be an list of 0-indexed integer labels of length n
-    """
-    accuracy = 0
-    k = max(truth) + 1 # number of labels
-    
-    # This is not optimal, but we're using small k, so it's no biggie.
-    for p in itertools.permutations(range(k)):
-        compare = [p[t] for t in truth]
-        accuracy = max(accuracy, metrics.accuracy_score(labels, compare))
-    
-    return accuracy
